room-thermostat/requirements.txt
paho-mqtt==1.4.0
pytest==5.2.1
PyYAML==5.1.2
stringcase==1.2.0